rename the file as well
