+++ src/components/BadgeLinkGeneratorDialog.vue
