+++ src/components/MonitorSettingDialog.vue
@@ -31,7 +31,7 @@
 
                     <button
                         class="btn btn-primary btn-add-group me-2"
-                        @click="$refs.badgeGeneratorDialog.show(monitor.id, monitor.name)"
+                        @click="$refs.badgeLinkGeneratorDialog.show(monitor.id, monitor.name)"
                     >
                         <font-awesome-icon icon="certificate" />
                         {{ $t("Open Badge Link Generator") }}
@@ -46,16 +46,16 @@
             </div>
         </div>
     </div>
-    <BadgeGeneratorDialog ref="badgeGeneratorDialog" />
+    <BadgeLinkGeneratorDialog ref="badgeLinkGeneratorDialog" />
 </template>
 
 <script lang="ts">
 import { Modal } from "bootstrap";
-import BadgeGeneratorDialog from "./BadgeGeneratorDialog.vue";
+import BadgeLinkGeneratorDialog from "./BadgeLinkGeneratorDialog.vue";
 
 export default {
     components: {
-        BadgeGeneratorDialog
+        BadgeLinkGeneratorDialog
     },
     props: {},
     emits: [],
